backend/client_services/account_service/requirements.txt
amqp==5.1.1
anyio==3.6.1
bcrypt==4.0.0
beanie==1.11.12
billiard==3.6.4.0
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
ecdsa==0.18.0
elastic-apm==6.13.2
fastapi==0.85.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
install==1.3.5
kombu==5.2.4
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
prompt-toolkit==3.0.31
pyasn1==0.4.8
pycrypto==2.6.1
pydantic==1.10.2
PyJWT==2.5.0
pymongo==4.2.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rsa==4.9
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.13
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
